refactor: streamline layer handling and enhance GeoTIFF processing capabilities
- Removed redundant console logs in various services for cleaner code.
- Introduced LayerDatabaseService and LayerProcessingService for improved database operations and file processing.
- Updated IPC handlers to utilize the new services for layer management and GeoTIFF processing.
- Added functionality to extract geographic bounds from GeoTIFF files and process them for display.
- Enhanced raster processing to handle GeoTIFF files and integrate bounds into layer metadata.
